+++ inc/Classes/MasterForm.php
@@ -611,7 +611,8 @@ class MasterForm
 
                                                 // Callbacks
                                                 } elseif ($field['callback']) {
-                                                      $err = call_user_func($field['callback'], $_POST[$field['name']]);
+                                                    $postFieldValue = $_POST[$field['name']] ?? '';
+                                                    $err = call_user_func($field['callback'], $postFieldValue);
                                                     if ($err) {
                                                         $this->error[$field['name']] = $err;
                                                     }
@@ -759,7 +760,8 @@ class MasterForm
                                                     ob_end_clean();
                                                     $dsp->AddSingleRow($fcke_content);
 
-                                                    if ($this->error[$field['name']]) {
+                                                    $errorMessage = $this->error[$field['name']] ?? '';
+                                                    if ($errorMessage) {
                                                         $dsp->AddDoubleRow($field['caption'], $dsp->errortext_prefix . $this->error[$field['name']] . $dsp->errortext_suffix);
                                                     }
                                                 } else {