+++ modules/faq/change_item.php
@@ -1,6 +1,7 @@
 <?php
 
-switch ($_GET["step"]) {
+$stepParameter = $_GET["step"] ?? 0;
+switch ($stepParameter ) {
     case 3:
         $get_cat_names = $db->qry("SELECT name FROM %prefix%faq_cat");
         while ($row=$db->fetch_array($get_cat_names)) {
@@ -45,7 +46,8 @@ switch ($_GET["step"]) {
         break;
 }
 
-switch ($_GET["step"]) {
+$stepParameter = $_GET["step"] ?? 0;
+switch ($stepParameter) {
     default:
         include('show.php');
         break;